Factor out base layer and heatmap into separate modules
