+++ js/src/jupyter-gmaps.js
@@ -4,6 +4,8 @@ import html2canvas from 'html2canvas'
 
 import GoogleMapsLoader from 'google-maps'
 
+import { GMapsLayerView, GMapsLayerModel } from './GMapsLayer';
+
 function reloadGoogleMaps(configuration) {
     GoogleMapsLoader.release();
     GoogleMapsLoader.LIBRARIES = ["visualization"] ;
@@ -36,13 +38,6 @@ const ConfigurationMixin = {
 
 
 // Views
-
-const GMapsLayerView = widgets.WidgetView.extend({
-    initialize(parameters) {
-        GMapsLayerView.__super__.initialize.apply(this, arguments)
-        this.mapView = this.options.mapView
-    }
-})
 
 
 export const DirectionsLayerView = GMapsLayerView.extend({
@@ -98,71 +93,6 @@ export const DirectionsLayerView = GMapsLayerView.extend({
     }
 })
 
-
-const HeatmapLayerBaseView = GMapsLayerView.extend({
-    render() {
-        this.modelEvents() ;
-        GoogleMapsLoader.load((google) => {
-            this.heatmap = new google.maps.visualization.HeatmapLayer({
-                data: this.getData(),
-                radius: this.model.get("point_radius"),
-                maxIntensity: this.model.get("max_intensity"),
-                dissipating: this.model.get("dissipating"),
-                opacity: this.model.get("opacity"),
-                gradient: this.model.get("gradient")
-            }) ;
-        });
-    },
-
-    addToMapView(mapView) {
-        this.heatmap.setMap(mapView.map)
-    },
-
-    modelEvents() {
-        // Simple properties:
-        // [nameInView, nameInModel]
-        const properties = [
-            ['maxIntensity', 'max_intensity'],
-            ['opacity', 'opacity'],
-            ['radius', 'point_radius'],
-            ['dissipating', 'dissipating'],
-            ['gradient', 'gradient']
-        ]
-        properties.forEach(([nameInView, nameInModel]) => {
-            const callback = (
-                () => this.heatmap.set(nameInView, this.model.get(nameInModel))
-            )
-            this.model.on(`change:${nameInModel}`, callback, this)
-        })
-    },
-
-    get_data() {},
-
-})
-
-export const SimpleHeatmapLayerView = HeatmapLayerBaseView.extend({
-    getData() {
-        const data = this.model.get("data")
-        const dataAsGoogle = new google.maps.MVCArray(
-            data.map(([lat, lng]) => new google.maps.LatLng(lat, lng))
-        )
-        return dataAsGoogle
-    }
-})
-
-
-export const WeightedHeatmapLayerView = HeatmapLayerBaseView.extend({
-    getData() {
-        const data = this.model.get("data")
-        const dataAsGoogle = new google.maps.MVCArray(
-            data.map(([lat, lng, weight]) => {
-                const location = new google.maps.LatLng(lat, lng)
-                return { location: location, weight: weight }
-            })
-        );
-        return dataAsGoogle
-    }
-})
 
 /* Base class for markers.
  * This sets options common to the different types of markers.
@@ -484,33 +414,10 @@ _.extend(PlainmapView.prototype, ConfigurationMixin);
 
 // Models
 
-export const GMapsLayerModel = widgets.WidgetModel.extend({
-    defaults: _.extend({}, widgets.WidgetModel.prototype.defaults, {
-        _view_name : 'GMapsLayerView',
-        _model_name : 'GMapsLayerModel',
-        _view_module : 'jupyter-gmaps',
-        _model_module : 'jupyter-gmaps'
-    })
-});
-
 export const DirectionsLayerModel = GMapsLayerModel.extend({
     defaults: _.extend({}, GMapsLayerModel.prototype.defaults, {
         _view_name: "DirectionsLayerView",
         _model_name: "DirectionsLayerModel"
-    })
-});
-
-export const SimpleHeatmapLayerModel = GMapsLayerModel.extend({
-    defaults: _.extend({}, GMapsLayerModel.prototype.defaults, {
-        _view_name: "SimpleHeatmapLayerView",
-        _model_name: "SimpleHeatmapLayerModel"
-    })
-});
-
-export const WeightedHeatmapLayerModel = GMapsLayerModel.extend({
-    defaults: _.extend({}, GMapsLayerModel.prototype.defaults, {
-        _view_name: "WeightedHeatmapLayerView",
-        _model_name: "WeightedHeatmapLayerModel"
     })
 });
